+++ lib/jei/error.rb
@@ -1,0 +1,4 @@
+module Jei
+  class Error < RuntimeError
+  end
+end